factor out cpu device
