+++ crabml-core/src/backends/cpu/mod.rs
@@ -1,6 +1,10 @@
 pub mod buf;
-pub mod cpu_tensor;
+mod cpu_device;
+mod cpu_tensor;
 mod primitives;
 
+pub use buf::CpuTensorBuf;
+pub use cpu_device::CpuTensorDevice;
+pub use cpu_device::CpuTensorDeviceOptions;
+pub use cpu_device::CpuTensorDeviceRef;
 pub use cpu_tensor::CpuTensor;
-pub use cpu_tensor::CpuTensorDeviceRef;
